context-cli/pyproject.toml
[project]
name = "context-cli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "pyyaml>=6.0.3",
]
